AGI: Add support for Mac Manhunter games
Macintosh versions of GR and MH1-2 use the same combined directory
filename convention as DOS, Atari ST and Apple IIgs (but not Amiga or
CoCo3). However, the filename convention for volumes uses the common
AGI v2 format, without a game id prefix.

Example:
DOS: GRDIR, GRVOL.x
Mac: GRDIR, VOL.x

To account for this, we'll use the following:
1. Rather than take the directory filename prefix from volume files,
   we now get this from the directory files themselves (which makes
   more sense anyway).
2. Later, when loading individual volumes, check for Mac platform and
   exclude the prefix from the path.

This improves the previous single game check for Mac Gold Rush, which
formed directory and volume paths for that game only. This check is
now obsolete and has been removed. The detection flag that was used to
invoke the workaround has also been removed.