DIRECTOR: Improve detection entries
Barbie Nail Designer was causing a detection conflict with JumpStart
Explorers due to generic filenames that were used. Adding the movie
path instead of relying on directory globs resolves this.

Alphabet Train was causing a detection conflict with a non-Director
file in JumpStart ABC's. Adding a Director movie as a second file
resolves this.