SCI: Add detection for Japanese Phant1 demo
Promotional release of chapter 1 in Japanese. The earlier Windows
version distributed by Sierra Pioneer was in English. The package
includes a coupon to redeem for the full version.

Since this includes only chapter 1 (and the menu allows selection of
later chapters), it is possible that this disc matches the files on
disc 1 of the full Japanese Windows version (SPRW-5001).